f3rm/features/orientany/requirements.txt
#Install manually
#torch

transformers
matplotlib
pillow
huggingface-hub
gradio
numpy==1.26.4
onnxruntime
rembg